move field components to registry
